requirements.txt
# Hallucination-Aware Adaptive LLM System
# ----------------------------------------
# Core
PyYAML==6.0.1
numpy==1.26.4

# ML / Prediction
scikit-learn==1.7.2
joblib==1.3.2

# NLP — Named Entity Recognition
spacy>=3.7.0
# After install, download the model: python -m spacy download en_core_web_sm

# NLI Verification
transformers>=4.40.0

# Embeddings & Vector Search
sentence-transformers==2.7.0
faiss-cpu==1.8.0

# LLM Integration
openai==1.30.1
tiktoken==0.7.0

# Demo UI
streamlit==1.35.0

# Testing
pytest==8.2.0
